refactor: extract dep reconciliation
